move embed.FS source to seperate file
for backwards compat with previous versions of the language the new code
is protected with a +build tag.
